+++ examples/dotnet/lambda-with-url-and-layers/dnapp/custom_lib_source_code/HelloWorld.cs
@@ -1,0 +1,10 @@
+namespace HelloWorldLibrary
+{
+    public class HelloWorld
+    {
+        public static string GetMessage()
+        {
+            return "Hello World from layer!";
+        }
+    }
+}