+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.14.1] - 2024-10-08
+- Added support lambda layers with runtime DotNet
+- Tagging added to `oas_v3 openapi` API Gateway.
+- SNS topic deletion fixed.
+- Added an example of a lambda function with runtime `dotnet`
+- Fixed silent overwriting existing lambda with runtime Java during the command `syndicate generate lambda`
+
 # [1.14.0] - 2024-08-28
 - Changed deployment flow to work despite the latest deployment failed
 - Changed deployment flow with the flag `--continue_deploy` to work despite the latest deployment being absent or succeeded